% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{ajouter_calories}
\alias{ajouter_calories}
\title{Ajouter les calories et identifier les observations hors plage}
\usage{
ajouter_calories(df, table_calories, table_menage)
}
\arguments{
\item{df}{Base avec QuantiteConsommeeKG et Produit}

\item{table_calories}{Table avec conversion en calories}

\item{table_menage}{Table avec IDs et membres du ménage}
}
\value{
Base avec variable CaloriesParTete et HorsPlageCalories
}
\description{
Ajouter les calories et identifier les observations hors plage
}
