% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{ajouter_contexte_menage}
\alias{ajouter_contexte_menage}
\title{Enrichir les données avec les variables de contexte (région, département, milieu)}
\usage{
ajouter_contexte_menage(df, base_all)
}
\arguments{
\item{df}{Base principale}

\item{base_all}{Base contextuelle avec s00q01 à s00q04}
}
\value{
Base enrichie
}
\description{
Enrichir les données avec les variables de contexte (région, département, milieu)
}
