% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{ajouter_quantite_kg}
\alias{ajouter_quantite_kg}
\title{Appliquer une table de conversion pour obtenir les quantités en kg}
\usage{
ajouter_quantite_kg(df, table_conversion)
}
\arguments{
\item{df}{Base de consommation avec Produit, unite, taille, QuantiteConsommee}

\item{table_conversion}{Table avec poids en grammes (colonnes : produitID, uniteID, tailleID, poids)}
}
\value{
Base avec variable QuantiteConsommeeKG et colonne Source
}
\description{
Appliquer une table de conversion pour obtenir les quantités en kg
}
