% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{ajouter_valeur_unitaire}
\alias{ajouter_valeur_unitaire}
\title{Calcul de la valeur unitaire (VU)}
\usage{
ajouter_valeur_unitaire(df)
}
\arguments{
\item{df}{Base contenant les variables de prix et quantités}
}
\value{
La base avec VU calculée et observations sans quantité filtrées
}
\description{
Calcul de la valeur unitaire (VU)
}
