% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{calculer_conso_par_source}
\alias{calculer_conso_par_source}
\title{Décomposer les consommations selon la source}
\usage{
calculer_conso_par_source(df, id_var)
}
\arguments{
\item{df}{Base avec VU et quantités par source}

\item{id_var}{Identifiant produit}
}
\value{
Base au format long : (source, quantité, valeur)
}
\description{
Décomposer les consommations selon la source
}
