% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{calculer_indicateurs_synthetiques}
\alias{calculer_indicateurs_synthetiques}
\title{Calculer les indicateurs synthétiques à partir d'une base appurée}
\usage{
calculer_indicateurs_synthetiques(
  base,
  base_x1 = NULL,
  var_source = "Source",
  var_produit = "Produit",
  var_achat = "dernier_achat_poisson"
)
}
\arguments{
\item{base}{Dataframe appuré (avec QuantiteConsommeeKG, ValeurConsommee, Calories, etc.)}

\item{base_x1}{(optionnel) Base brute avec une variable "dernier_achat_poisson" pour l'indicateur 9}

\item{var_source}{Nom de la variable source ("Source" par défaut)}

\item{var_produit}{Nom de la variable produit ("Produit" par défaut)}

\item{var_achat}{Nom de la variable d'achat récent dans la base brute ("dernier_achat_poisson")}
}
\value{
Liste avec les 10 indicateurs calculés
}
\description{
Calculer les indicateurs synthétiques à partir d'une base appurée
}
