% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{calculer_valeurs_consommees}
\alias{calculer_valeurs_consommees}
\title{Application des valeurs unitaires pour obtenir la consommation monétaire}
\usage{
calculer_valeurs_consommees(df, baseVU, id_var)
}
\arguments{
\item{df}{Base d'origine (après ajout des VU individuelles)}

\item{baseVU}{Base des valeurs unitaires (produit, unité, taille, VU)}

\item{id_var}{Identifiant du produit (ex: "poissons__id")}
}
\value{
Base contenant les dépenses estimées (avec VU_ref)
}
\description{
Application des valeurs unitaires pour obtenir la consommation monétaire
}
