% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{construire_base_VU}
\alias{construire_base_VU}
\title{Construction de la base de référence des valeurs unitaires (VU)}
\usage{
construire_base_VU(df, id_var)
}
\arguments{
\item{df}{Base avec VU et identifiants}

\item{id_var}{Nom de l'identifiant produit (ex: "poissons__id")}
}
\value{
Base VU (produit, unité, taille, VU représentatif)
}
\description{
Construction de la base de référence des valeurs unitaires (VU)
}
