% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{exporter_autres_produits}
\alias{exporter_autres_produits}
\title{Exporter les autres produits non listés}
\usage{
exporter_autres_produits(df, nom_base)
}
\arguments{
\item{df}{Dataframe brut}

\item{nom_base}{Nom de la base (ex: "poissons")}
}
\value{
Dataframe des autres produits avec leurs descriptions
}
\description{
Exporter les autres produits non listés
}
