% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\docType{package}
\name{premierpackage}
\alias{-package}
\alias{premierpackage}
\title{Traitement des données alimentaires EHCVM}
\description{
Fonctions génériques pour traiter les fichiers de consommation
alimentaire (poissons, viandes, céréales, etc.) de l'enquête EHCVM.
}
