% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{prepare_base}
\alias{prepare_base}
\title{Nettoyage et préparation de la base alimentaire}
\usage{
prepare_base(df, nom_base, id_var)
}
\arguments{
\item{df}{Dataframe brut lu avec haven::read_dta}

\item{nom_base}{Nom de la base (ex: "poissons", "viandes")}

\item{id_var}{Nom de l'identifiant produit (ex: "poissons__id")}
}
\value{
Dataframe nettoyé avec variables renommées, labellisées et cohérentes
}
\description{
Nettoyage et préparation de la base alimentaire
}
