% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement.R
\name{winsoriser_quantite}
\alias{winsoriser_quantite}
\title{Winsoriser une variable quantitative groupée par modalité}
\usage{
winsoriser_quantite(df)
}
\arguments{
\item{df}{Base avec QuantiteConsommee et variable Source}
}
\value{
Base avec colonne QuantiteConsommee1 winsorisée et IDs disponibles
}
\description{
Winsoriser une variable quantitative groupée par modalité
}
